language-models/langchain_ng_aoai/apis/requirements.txt
#Copyright (c) Microsoft. All rights reserved.
python-dotenv==1.0.0
azure-storage-blob==12.14.1
requests==2.28.2
tiktoken==0.2.0
langchain==0.0.271
azure-search-documents==11.4.0b8
fastapi==0.109.0
uvicorn[standard]==0.26.0
gunicorn==21.2.0
pandas==1.5.3
azure-monitor-opentelemetry==1.3.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
azure-search-documents==11.5.1